test: Enable running tests with Ganache + test refactor + use gentx for Gravity keys (#125)
This PR enables the use of the flag `PEGGO_E2E_USE_GANACHE=true` to run tests against a Ganache instance instead of Geth. This allows (and encourages) devs to run the E2E locally and more often. **It shouldn't be used in the CI.**

It's basically scratching my own itch of wanting to run E2E tests as fast as possible 😄